Elegant confirm with prompt

+++ dynamo/system.lua
@@ -1,70 +1,71 @@
 -- vim:filetype=lua:expandtab:shiftwidth=4:tabstop=8:softtabstop=4:fdm=marker:foldmarker={{{,}}}
 local html = require("html")
--- {{{ Lock function when using LightDM
-dynamo.lock = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Khóa màn hình (gõ 'Enter' để xác nhận)? ") },
+
+local function confirm_prompt(text_prompt, callback, text_confirm)
+    local text_confirm = text_confirm or "Gõ 'Enter' để xác nhận? "
+    dynamo.prompt(text_prompt)
+    awful.prompt.run({ prompt = html(beautiful.fg_confirm, text_confirm) },
                      mypromptbox[mouse.screen].widget,
                      function (t)
-                         awful.util.spawn("dm-tool switch-to-greeter")
+                         dynamo.prompt()
+                         callback(t)
+                     end,
+                     nil, nil, nil,
+                     function (t)
+                         dynamo.prompt()
                      end)
+end
+-- {{{ Lock function when using LightDM
+dynamo.lock = function()
+    confirm_prompt(" Khóa màn hình", function()
+        awful.util.spawn("dm-tool switch-to-greeter")
+    end)
 end
 -- }}}
 
 -- {{{ Quit function when using GNOME session
 dynamo.quit = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Thoát phiên (gõ 'Enter' để xác nhận)? ") },
-                     mypromptbox[mouse.screen].widget,
-                     function (t)
-                         awesome.emit_signal("exit", nil)
-                         if os.getenv("DESKTOP_SESSION") == "awesome-gnome" then
-                             awful.util.spawn("gnome-session-quit --no-prompt")
-                         else
-                             awesome.quit()
-                         end
-                     end)
+    confirm_prompt(" Thoát phiên", function ()
+        awesome.emit_signal("exit", nil)
+        if os.getenv("DESKTOP_SESSION") == "awesome-gnome" then
+            awful.util.spawn("gnome-session-quit --no-prompt")
+        else
+            awesome.quit()
+        end
+    end)
 end
 -- }}}
 
 -- {{{ Hibernate using systemd
 dynamo.hibernate = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Ngủ đông (gõ 'Enter' để xác nhận)? ") },
-                     mypromptbox[mouse.screen].widget,
-                     function (t)
-                         awful.util.spawn("systemctl hibernate")
-                     end)
+    confirm_prompt(" Ngủ đông", function ()
+        awful.util.spawn("systemctl hibernate")
+    end)
 end
 -- }}}
 
 -- {{{ Reboot
 dynamo.reboot = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Khởi động lại (gõ 'Enter' để xác nhận)? ") },
-                     mypromptbox[mouse.screen].widget,
-                     function (t)
-                         awesome.emit_signal("exit", nil)
-                         awful.util.spawn("reboot")
-                     end)
+    confirm_prompt(" Khởi động lại", function ()
+        awesome.emit_signal("exit", nil)
+        awful.util.spawn("reboot")
+    end)
 end
 -- }}}
 
 -- {{{ Shutdown
 dynamo.shutdown = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Tắt máy (gõ 'Enter' để xác nhận)? ") },
-                     mypromptbox[mouse.screen].widget,
-                     function (t)
-                         awesome.emit_signal("exit", nil)
-                         awful.util.spawn("shutdown -P now")
-                     end)
+    confirm_prompt(" Tắt máy", function ()
+        awesome.emit_signal("exit", nil)
+        awful.util.spawn("shutdown -P now")
+    end)
 end
 -- }}}
 
 -- {{{ Shutdown when at a time
 dynamo.shutdown_schedule = function()
-    awful.prompt.run({ prompt = html(beautiful.fg_confirm, "Tắt máy (nhập vào thời gian)? ") },
-                     mypromptbox[mouse.screen].widget,
-                     function (t)
-                         if string.lower(t) == 'y' or string.lower(t) == 'c' then
-                             awful.util.spawn("shutdown -P " .. trim(t))
-                         end
-                     end)
+    confirm_prompt(" Tắt máy", function (t)
+        awful.util.spawn("shutdown -P " .. trim(t))
+    end,"Nhập vào thời gian ")
 end
 -- }}}
